refactor(lyrics): split genius.rs tests into sibling file
genius.rs hit exactly 1000/1000 lines after recent work — any future
edit would break the airuleset cap. Extracted the test module to a
sibling file via #[path] (same pattern as PR #106 worker_tests/mod_tests
and #104 orchestrator_tests):

- genius.rs is now 448 lines (was 1000)
- genius_tests.rs is 555 lines

No behavior change; existing tests run unchanged from the new location.

Closes #95 (genius.rs portion — playback/mod.rs split tracked separately)
